Vulnerabilities/2020/CVE-2020-1377/prprhyt/requirements.txt
cryptography==2.9.2
pycryptodome==3.9.7
scapy==2.4.3
